Fix for mobilesafari
Weird quirk where a:link, a:visited { &:not(nav *) { &::after had :is() injected to shorten the selector, which should be safe (and is in other browsers) but was failing *for the ::after specifically*. Adding :not to each individual a: removes the :is in this scenario, and now it works. I think that's a MobileSafari bug, but who knows.